Migrate check-charset.py for fontbakery 0.12
